requirements.txt
Django
djangorestframework
python-dotenv
supabase
psycopg2-binary
python-dateutil
environs